requirements.txt
alembic==1.8.0
amqp==2.6.1
async-timeout==4.0.2
billiard==3.6.4.0
cachelib==0.9.0
celery==4.4.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
croniter==1.3.5
Deprecated==1.2.13
Flask==2.1.2
Flask-Caching==2.0.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
importlib-metadata==4.12.0
importlib-resources==5.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==4.6.11
Mako==1.2.1
MarkupSafe==2.1.1
packaging==21.3
prompt-toolkit==3.0.30
py3dns==3.2.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.39
validate-email==1.3
vine==1.3.0
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
